# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "A BCF API"
  version: "1.0.0"
  title: "BCFAPI"
host: "virtserver.swaggerhub.com"
basePath: "/swisx/bcfAPI/1.0.0"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /bcf/{version}/projects/{project_id}/topics:
    get:
      tags:
      - "topic"
      summary: "Retrieve a collection of topics related to a project"
      description: "Retrieve a collection of topics related to a project (default\
        \ sort order is creation_date)."
      operationId: "bcfVersionProjectsProject_idTopicsGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/topic_GET"
          examples:
            application/json:
            - guid: "A245F4F2-2C01-B43B-B612-5E456BEF8116"
              creation_author: "Architect@example.com"
              title: "Example topic 1"
              labels:
              - "Architecture"
              - "Structural"
              creation_date: "2013-10-21T17:34:22.409Z"
            - guid: "A211FCC2-3A3B-EAA4-C321-DE22ABC8414"
              creation_author: "Architect@example.com"
              title: "Example topic 2"
              labels:
              - "Architecture"
              - "Heating"
              - "Electrical"
              creation_date: "2014-11-19T14:24:11.316Z"
      x-swagger-router-controller: "Topic"
    post:
      tags:
      - "topic"
      summary: "Add a new topic."
      description: "Add a new topic. This operation is only possible when the server\
        \ returns the createTopic flag in the Project authorization."
      operationId: "bcfVersionProjectsProject_idTopicsPOST"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - in: "body"
        name: "Topic_POST"
        required: false
        schema:
          $ref: "#/definitions/Topic_POST"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/topic_GET"
          examples:
            application/json:
              guid: "A245F4F2-2C01-B43B-B612-5E456BEF8116"
              creation_author: "Architect@example.com"
              creation_date: "2016-08-01T17:34:22.409Z"
              topic_type: "Clash"
              topic_status: "open"
              title: "Example topic 3"
              priority: "high"
              labels:
              - "Architecture"
              - "Heating"
              assigned_to: "harry.muster@example.com"
              bim_snippet:
                snippet_type: "clash"
                is_external: true
                reference: "https://example.com/bcf/1.0/ADFE23AA11BCFF444122BB"
                reference_schema: "https://example.com/bcf/1.0/clash.xsd"
      x-swagger-router-controller: "Topic"
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}:
    get:
      tags:
      - "topic"
      summary: "Retrieve a specific topic."
      description: "Retrieve a specific topic."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/topic_GET"
          examples:
            application/json:
              guid: "B345F4F2-3A04-B43B-A713-5E456BEF8228"
              creation_author: "Architect@example.com"
              creation_date: "2016-08-01T17:34:22.409Z"
              topic_type: "Clash"
              topic_status: "open"
              title: "Example topic 3"
              priority: "high"
              labels:
              - "Architecture"
              - "Heating"
              assigned_to: "harry.muster@example.com"
              bim_snippet:
                snippet_type: "clash"
                is_external: true
                reference: "https://example.com/bcf/1.0/ADFE23AA11BCFF444122BB"
                reference_schema: "https://example.com/bcf/1.0/clash.xsd"
              authorization:
                topic_actions:
                - "createComment"
                - "createViewpoint"
      x-swagger-router-controller: "Topic"
    put:
      tags:
      - "topic"
      summary: "Modify a specific topic"
      description: "Modify a specific topic, description similar to POST. This operation\
        \ is only possible when the server returns the update flag in the Topic authorization"
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidPUT"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - in: "body"
        name: "Topic_PUT"
        required: false
        schema:
          $ref: "#/definitions/Topic_PUT"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/topic_GET"
          examples:
            application/json: null
      x-swagger-router-controller: "Topic"
    delete:
      tags:
      - "topic"
      summary: "Deletes a single topic."
      description: "Deletes a single topic. This operation is only possible when the\
        \ server returns the delete flag in the Topic authorization."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidDELETE"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Topic"
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}/snippet:
    get:
      tags:
      - "topic"
      summary: "Retrieves a topics BIM-Snippet as binary file."
      description: "Retrieves a topics BIM-Snippet as binary file."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidSnippetGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/bim_snippet"
          examples:
            application/json:
              bim_snippet:
                snippet_type: "clash"
                is_external: true
                reference: "https://example.com/bcf/1.0/ADFE23AA11BCFF444122BB"
                reference_schema: "https://example.com/bcf/1.0/clash.xsd"
      x-swagger-router-controller: "Topic"
    put:
      tags:
      - "topic"
      summary: "Puts a new BIM Snippet binary file to a topic"
      description: "Puts a new BIM Snippet binary file to a topic. If this is used,\
        \ the parent topics BIM Snippet property is_external must be set to false\
        \ and the reference must be the file name with extension. This operation is\
        \ only possible when the server returns the updateBimSnippet flag in the Topic\
        \ authorization. THERE IS NO BODY EXAMPLE!"
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidSnippetPUT"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Topic"
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}/files:
    get:
      tags:
      - "file"
      summary: "Retrieve a collection of file references as topic header."
      description: "Retrieve a collection of file references as topic header."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidFilesGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/file_GET"
          examples:
            application/json:
            - ifc_project: "0J$yPqHBD12v72y4qF6XcD"
              file_name: "OfficeBuilding_Architecture_0001.ifc"
              reference: "https://example.com/files/0J$yPqHBD12v72y4qF6XcD_0001.ifc"
            - ifc_project: "3hwBHP91jBRwPsmyf$3Hea"
              file_name: "OfficeBuilding_Heating_0003.ifc"
              reference: "cf37bae6-0900-46be-b37f-b34754fe0b4a"
      x-swagger-router-controller: "File"
    put:
      tags:
      - "file"
      summary: "Update a collection of file references on the topic header."
      description: "Update a collection of file references on the topic header. This\
        \ operation is only possible when the server returns the updateFiles flag\
        \ in the Topic authorization."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidFilesPUT"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - in: "body"
        name: "File_PUT"
        required: false
        schema:
          $ref: "#/definitions/File_PUT"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/file_GET"
          examples:
            application/json:
            - ifc_project: "0J$yPqHBD12v72y4qF6XcD"
              file_name: "OfficeBuilding_Architecture_0001.ifc"
              reference: "https://example.com/files/0J$yPqHBD12v72y4qF6XcD_0001.ifc"
            - ifc_project: "3hwBHP91jBRwPsmyf$3Hea"
              file_name: "OfficeBuilding_Heating_0003.ifc"
              reference: "https://example.com/files/3hwBHP91jBRwPsmyf$3Hea_0003.ifc"
      x-swagger-router-controller: "File"
...
  /bcf/{version}/projects/{project_id}/topics/{topic_guid}/related_topics:
    get:
      tags:
      - "related topic"
      summary: "Retrieve a collection of all related topics to a topic."
      description: "Retrieve a collection of all related topics to a topic."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidRelated_topicsGET"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/related_topic_GET"
          examples:
            application/json:
            - related_topic_guid: "db49df2b-0e42-473b-a3ee-f7b785d783c4"
            - related_topic_guid: "6963a846-54d1-4050-954d-607cd5e48aa3"
      x-swagger-router-controller: "RelatedTopic"
    put:
      tags:
      - "related topic"
      summary: "Add or update a collection of all related topics to a topic."
      description: "Add or update a collection of all related topics to a topic. This\
        \ operation is only possible when the server returns the updateRelatedTopics\
        \ flag in the Topic authorization."
      operationId: "bcfVersionProjectsProject_idTopicsTopic_guidRelated_topicsPUT"
      parameters:
      - name: "version"
        in: "path"
        description: "BFC Version"
        required: true
        type: "string"
        enum:
        - "1.0"
        - "2.1"
      - name: "project_id"
        in: "path"
        description: "Project ID"
        required: true
        type: "string"
      - name: "topic_guid"
        in: "path"
        description: "Topic guid"
        required: true
        type: "string"
      - in: "body"
        name: "related_topic_PUT"
        required: false
        schema:
          $ref: "#/definitions/related_topic_GET"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/related_topic_GET"
          examples:
            application/json:
            - related_topic_guid: "db49df2b-0e42-473b-a3ee-f7b785d783c4"
            - related_topic_guid: "6963a846-54d1-4050-954d-607cd5e48aa3"
            - related_topic_guid: "bac66ab4-331e-4f21-a28e-083d2cf2e796"
      x-swagger-router-controller: "RelatedTopic"
...
definitions:
...
  file_GET:
    type: "object"
    properties:
      ifc_project:
        type: "string"
      ifc_spatial_structure_element:
        type: "string"
      file_name:
        type: "string"
      date:
        type: "string"
      reference:
        type: "string"
    description: "Schema for a single file GET, BCF REST API."
  related_topic_GET:
    type: "object"
    required:
    - "related_topic_guid"
    properties:
      related_topic_guid:
        type: "string"
    description: "Schema for single related topic GET, BCF REST API."
    example:
    - related_topic_guid: "db49df2b-0e42-473b-a3ee-f7b785d783c4"
    - related_topic_guid: "6963a846-54d1-4050-954d-607cd5e48aa3"
    - related_topic_guid: "bac66ab4-331e-4f21-a28e-083d2cf2e796"
...
  topic_GET:
    type: "object"
    properties:
      guid:
        type: "string"
      topic_type:
        type: "string"
      topic_status:
        type: "string"
      reference_links:
        type: "array"
        items:
          type: "string"
      title:
        type: "string"
      priority:
        type: "string"
      index:
        type: "integer"
      labels:
        type: "array"
        items:
          type: "string"
      creation_date:
        type: "string"
      creation_author:
        type: "string"
      modified_date:
        type: "string"
      modified_author:
        type: "string"
      assigned_to:
        type: "string"
      stage:
        type: "string"
      description:
        type: "string"
      bim_snippet:
        $ref: "#/definitions/bim_snippet"
      due_date:
        type: "string"
      authorization:
        $ref: "#/definitions/topic_GET_authorization"
...
  bim_snippet:
    type: "object"
    properties:
      snippet_type:
        type: "string"
      is_external:
        type: "boolean"
      reference:
        type: "string"
      reference_schema:
        type: "string"
  topic_actions:
    type: "array"
    items:
      type: "string"
      example: ""
...
  Topic_POST:
    type: "object"
    properties:
      guid:
        type: "string"
      topic_type:
        type: "string"
      topic_status:
        type: "string"
      reference_links:
        type: "array"
        items:
          type: "string"
      title:
        type: "string"
      priority:
        type: "string"
      index:
        type: "integer"
      labels:
        type: "array"
        items:
          type: "string"
      creation_date:
        type: "string"
      creation_author:
        type: "string"
      modified_date:
        type: "string"
      modified_author:
        type: "string"
      assigned_to:
        type: "string"
      stage:
        type: "string"
      description:
        type: "string"
      bim_snippet:
        $ref: "#/definitions/bim_snippet"
      due_date:
        type: "string"
    example:
      topic_type: "Clash"
      topic_status: "open"
      title: "Example topic 3"
      priority: "high"
      labels:
      - "Architecture"
      - "Heating"
      assigned_to: "harry.muster@example.com"
      bim_snippet:
        snippet_type: "clash"
        is_external: true
        reference: "https://example.com/bcf/1.0/ADFE23AA11BCFF444122BB"
        reference_schema: "https://example.com/bcf/1.0/clash.xsd"
  Topic_PUT:
    type: "object"
    properties:
      guid:
        type: "string"
      topic_type:
        type: "string"
      topic_status:
        type: "string"
      reference_links:
        type: "array"
        items:
          type: "string"
      title:
        type: "string"
      priority:
        type: "string"
      index:
        type: "integer"
      labels:
        type: "array"
        items:
          type: "string"
      creation_date:
        type: "string"
      creation_author:
        type: "string"
      modified_date:
        type: "string"
      modified_author:
        type: "string"
      assigned_to:
        type: "string"
      stage:
        type: "string"
      description:
        type: "string"
      bim_snippet:
        $ref: "#/definitions/bim_snippet"
      due_date:
        type: "string"
    example:
      topic_type: "Clash"
      topic_status: "open"
      title: "Example topic 3 - Changed Title"
      priority: "high"
      labels:
      - "Architecture"
      - "Heating"
      assigned_to: "harry.muster@example.com"
      bim_snippet:
        snippet_type: "clash"
        is_external: true
        reference: "https://example.com/bcf/1.0/ADFE23AA11BCFF444122BB"
        reference_schema: "https://example.com/bcf/1.0/clash.xsd"
  File_PUT:
    type: "object"
    properties:
      ifc_project:
        type: "string"
      ifc_spatial_structure_element:
        type: "string"
      file_name:
        type: "string"
      date:
        type: "string"
      reference:
        type: "string"
    example:
      ifc_project: "0J$yPqHBD12v72y4qF6XcD"
      file_name: "OfficeBuilding_Architecture_0001.ifc"
      reference: "https://example.com/files/0J$yPqHBD12v72y4qF6XcD_0001.ifc"
...
  topic_GET_authorization:
    properties:
      topic_actions:
        $ref: "#/definitions/topic_actions"
      topic_status:
        type: "array"
        items:
          type: "string"
...